Fix typo in comment. Raise embedded if into parent if.

+++ commons-vfs2/src/test/java/org/apache/commons/vfs2/impl/DefaultFileMonitorTest.java
@@ -69,10 +69,8 @@ public class DefaultFileMonitorTest {
 
     @After
     public void tearDown() throws Exception {
-        if (testFile != null) {
-            if (!testFile.delete()) {
-                testFile.deleteOnExit();
-            }
+        if (testFile != null && !testFile.delete()) {
+            testFile.deleteOnExit();
         }
     }
